Support Keyman10 Keyboard Switching
* Refactor KeyboardSwitchingAdapter
* Implement new adapter for KM 10
