+++ SIL.Windows.Forms.Keyboarding/Windows/WinKeyboardAdaptor.cs
@@ -1,16 +1,13 @@
-// Copyright (c) 2013 SIL International
+// Copyright (c) 2013-2018 SIL International
 // This software is licensed under the MIT License (http://opensource.org/licenses/MIT)
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Runtime.InteropServices;
-using System.Text;
 using System.Windows.Forms;
-using Microsoft.Win32;
 using SIL.Keyboarding;
 
 namespace SIL.Windows.Forms.Keyboarding.Windows
@@ -18,30 +15,8 @@ namespace SIL.Windows.Forms.Keyboarding.Windows
 	/// <summary>
 	/// Class for handling Windows system keyboards
 	/// </summary>
-	internal class WinKeyboardAdaptor : IKeyboardRetrievingAdaptor, IKeyboardSwitchingAdaptor
+	internal class WinKeyboardAdaptor : IKeyboardRetrievingAdaptor
 	{
-		internal class LayoutName
-		{
-			public LayoutName()
-			{
-				Name = string.Empty;
-				LocalizedName = string.Empty;
-			}
-
-			public LayoutName(string layout): this(layout, layout)
-			{
-			}
-
-			public LayoutName(string layout, string localizedLayout)
-			{
-				Name = layout;
-				LocalizedName = localizedLayout;
-			}
-
-			public string Name;
-			public string LocalizedName;
-		}
-
 		/// <summary>
 		/// This class receives notifications from TSF when the input method changes.
 		/// It also implements a fallback to Windows messages if TSF isn't available,
@@ -70,7 +45,7 @@ namespace SIL.Windows.Forms.Keyboarding.Windows
 
 			public void OnLanguageChanged()
 			{
-				IKeyboardDefinition winKeyboard = _keyboardAdaptor.ActiveKeyboard;
+				IKeyboardDefinition winKeyboard = _keyboardAdaptor.SwitchingAdaptor.ActiveKeyboard;
 				Debug.WriteLine("Language changed from {0} to {1}",
 					Keyboard.Controller.ActiveKeyboard != null ? Keyboard.Controller.ActiveKeyboard.Layout : "<null>",
 					winKeyboard != null ? winKeyboard.Layout : "<null>");
@@ -122,11 +97,8 @@ namespace SIL.Windows.Forms.Keyboarding.Windows
 
 		}
 
-		private Timer _timer;
 		private WinKeyboardDescription _expectedKeyboard;
 		private bool _fSwitchedLanguages;
-		/// <summary>Used to prevent re-entrancy. <c>true</c> while we're in the middle of switching keyboards.</summary>
-		private bool _fSwitchingKeyboards;
 
 		private ushort _profileNotifySinkCookie;
 		private readonly TfLanguageProfileNotifySink _tfLanguageProfileNotifySink;
@@ -136,6 +108,10 @@ namespace SIL.Windows.Forms.Keyboarding.Windows
 		internal ITfInputProcessorProfiles ProcessorProfiles { get; private set; }
 		internal ITfInputProcessorProfileMgr ProfileMgr { get; private set; }
 		internal ITfSource TfSource { get; private set; }
+
+		public bool IsApplicable => true;
+
+		public IKeyboardSwitchingAdaptor SwitchingAdaptor { get; private set; }
 
 		public WinKeyboardAdaptor()
 		{
@@ -152,6 +128,7 @@ namespace SIL.Windows.Forms.Keyboarding.Windows
 			// ProfileMgr will be null on Windows XP - the interface got introduced in Vista
 			ProfileMgr = ProcessorProfiles as ITfInputProcessorProfileMgr;
 
+			SwitchingAdaptor = new WindowsXPKeyboardSwitchingAdapter(ProcessorProfiles, ProfileMgr);
 			_tfLanguageProfileNotifySink = new TfLanguageProfileNotifySink(this);
 
 			TfSource = ProcessorProfiles as ITfSource;
@@ -288,7 +265,7 @@ namespace SIL.Windows.Forms.Keyboarding.Windows
 			return string.Format("{0} - {1}", layout, locale);
 		}
 
-		private void GetInputMethods()
+		public void UpdateAvailableKeyboards()
 		{
 			IEnumerable<Tuple<TfInputProcessorProfile, ushort, IntPtr>> imes;
 			if (ProfileMgr != null)
@@ -334,7 +311,7 @@ namespace SIL.Windows.Forms.Keyboarding.Windows
 					}
 					else
 					{
-						layoutName = GetLayoutNameEx(hkl);
+						layoutName = WinKeyboardUtils.GetLayoutNameEx(hkl);
 					}
 
 					string id = GetId(layoutName.Name, locale);
@@ -372,346 +349,13 @@ namespace SIL.Windows.Forms.Keyboarding.Windows
 					// Keyman keyboard (LT-16172)
 				}
 			}
-
-			foreach (WinKeyboardDescription existingKeyboard in curKeyboards.Values)
+			// Set each unhanandled keyboard to unavailable
+			foreach (var existingKeyboard in curKeyboards.Values)
+			{
 				existingKeyboard.SetIsAvailable(false);
-
-		}
-
-		internal static LayoutName GetLayoutNameEx(IntPtr handle)
-		{
-			// InputLanguage.LayoutName is not to be trusted, especially where there are mutiple
-			// layouts (input methods) associated with a language. This function also provides
-			// the additional benefit that it does not matter whether a user switches from using
-			// InKey in Portable mode to using it in Installed mode (perhaps as the project is
-			// moved from one computer to another), as this function will identify the correct
-			// input language regardless, rather than (unhelpfully ) calling an InKey layout in
-			// portable mode the "US" layout. The layout is identified soley by the high-word of
-			// the HKL (a.k.a. InputLanguage.Handle).  (The low word of the HKL identifies the
-			// language.)
-			// This function determines an HKL's LayoutName based on the following order of
-			// precedence:
-			// - Look up HKL in HKCU\\Software\\InKey\\SubstituteLayoutNames
-			// - Look up extended layout in HKLM\\SYSTEM\\CurrentControlSet\\Control\\Keyboard Layouts
-			// - Look up basic (non-extended) layout in HKLM\\SYSTEM\\CurrentControlSet\\Control\\Keyboard Layouts
-			// -Scan for ID of extended layout in HKLM\\SYSTEM\\CurrentControlSet\\Control\\Keyboard Layouts
-			var hkl = string.Format("{0:X8}", (long)handle);
-
-			// Get substitute first
-			var substituteHkl = (string)Registry.GetValue(@"HKEY_CURRENT_USER\Keyboard Layout\Substitutes", hkl, null);
-			if (!string.IsNullOrEmpty(substituteHkl))
-				hkl = substituteHkl;
-
-			// Check InKey
-			var substituteLayoutName = (string)Registry.GetValue(@"HKEY_CURRENT_USER\Software\InKey\SubstituteLayoutNames", hkl, null);
-			if (!string.IsNullOrEmpty(substituteLayoutName))
-				return new LayoutName(substituteLayoutName);
-
-			var layoutName = GetLayoutNameFromKey(string.Concat(@"HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Control\Keyboard Layouts\", hkl));
-			if (layoutName != null)
-				return layoutName;
-
-			layoutName = GetLayoutNameFromKey(string.Concat(@"HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Control\Keyboard Layouts\0000",
-				hkl.Substring(0, 4)));
-			if (layoutName != null)
-				return layoutName;
-
-			using (var regKey = Registry.LocalMachine.OpenSubKey(@"SYSTEM\CurrentControlSet\Control\Keyboard Layouts"))
-			{
-				if (regKey == null)
-					return new LayoutName();
-
-				string layoutId = "0" + hkl.Substring(1, 3);
-				foreach (string subKeyName in regKey.GetSubKeyNames().Reverse())
-					// Scan in reverse order for efficiency, as the extended layouts are at the end.
-				{
-					using (var klid = regKey.OpenSubKey(subKeyName))
-					{
-						if (klid == null)
-							continue;
-
-						var layoutIdSk = ((string) klid.GetValue("Layout ID"));
-						if (layoutIdSk != null &&
-							layoutIdSk.Equals(layoutId, StringComparison.InvariantCultureIgnoreCase))
-						{
-							return GetLayoutNameFromKey(klid.Name);
-						}
-					}
-				}
-			}
-
-			return new LayoutName();
-		}
-
-		private static LayoutName GetLayoutNameFromKey(string key)
-		{
-			var layoutText = (string)Registry.GetValue(key, "Layout Text", null);
-			var displayName = (string)Registry.GetValue(key, "Layout Display Name", null);
-			if (string.IsNullOrEmpty(layoutText) && string.IsNullOrEmpty(displayName))
-				return null;
-
-			if (string.IsNullOrEmpty(displayName))
-				return new LayoutName(layoutText);
-
-			var bldr = new StringBuilder(100);
-			Win32.SHLoadIndirectString(displayName, bldr, 100, IntPtr.Zero);
-			return string.IsNullOrEmpty(layoutText) ? new LayoutName(bldr.ToString()) :
-				new LayoutName(layoutText, bldr.ToString());
-		}
-
-		/// <summary>
-		/// Gets the InputLanguage that has the same layout as <paramref name="keyboardDescription"/>.
-		/// </summary>
-		internal static InputLanguage GetInputLanguage(WinKeyboardDescription keyboardDescription)
-		{
-			InputLanguage sameLayout = null;
-			InputLanguage sameCulture = null;
-			foreach (InputLanguage lang in InputLanguage.InstalledInputLanguages)
-			{
-				// TODO: write some tests
-				try
-				{
-					if (GetLayoutNameEx(lang.Handle).Name == keyboardDescription.Name)
-					{
-						if (keyboardDescription.Locale == lang.Culture.Name)
-							return lang;
-						if (sameLayout == null)
-							sameLayout = lang;
-					}
-					else if (keyboardDescription.Locale == lang.Culture.Name && sameCulture == null)
-						sameCulture = lang;
-				}
-				catch (CultureNotFoundException)
-				{
-					// we get an exception for non-supported cultures, probably because of a
-					// badly applied .NET patch.
-					// http://www.ironspeed.com/Designer/3.2.4/WebHelp/Part_VI/Culture_ID__XXX__is_not_a_supported_culture.htm and others
-				}
-			}
-			return sameLayout ?? sameCulture;
-		}
-
-		/// <summary>
-		/// Gets the keyboard description for the layout of <paramref name="inputLanguage"/>.
-		/// </summary>
-		private static KeyboardDescription GetKeyboardDescription(IInputLanguage inputLanguage)
-		{
-			KeyboardDescription sameLayout = KeyboardController.NullKeyboard;
-			KeyboardDescription sameCulture = KeyboardController.NullKeyboard;
-			// TODO: write some tests
-			string requestedLayout = GetLayoutNameEx(inputLanguage.Handle).Name;
-			foreach (WinKeyboardDescription keyboardDescription in KeyboardController.Instance.AvailableKeyboards.OfType<WinKeyboardDescription>())
-			{
-				try
-				{
-					if (requestedLayout == keyboardDescription.Layout)
-					{
-						if (keyboardDescription.Locale == inputLanguage.Culture.Name)
-							return keyboardDescription;
-						if (sameLayout == null)
-							sameLayout = keyboardDescription;
-					}
-					else if (keyboardDescription.Locale == inputLanguage.Culture.Name && sameCulture == null)
-						sameCulture = keyboardDescription;
-				}
-				catch (CultureNotFoundException)
-				{
-					// we get an exception for non-supported cultures, probably because of a
-					// badly applied .NET patch.
-					// http://www.ironspeed.com/Designer/3.2.4/WebHelp/Part_VI/Culture_ID__XXX__is_not_a_supported_culture.htm and others
-				}
-			}
-			return sameLayout ?? sameCulture;
-		}
-
-		private void OnTimerTick(object sender, EventArgs eventArgs)
-		{
-			if (_expectedKeyboard == null || !_fSwitchedLanguages)
-				return;
-
-			// This code gets only called if TSF is not available(e.g. Windows XP)
-			if (InputLanguage.CurrentInputLanguage.Culture.KeyboardLayoutId == _expectedKeyboard.InputLanguage.Culture.KeyboardLayoutId)
-			{
-				_expectedKeyboard = null;
-				_fSwitchedLanguages = false;
-				_timer.Enabled = false;
-				return;
-			}
-
-			SwitchKeyboard(_expectedKeyboard);
-		}
-
-		private bool UseWindowsApiForKeyboardSwitching(WinKeyboardDescription winKeyboard)
-		{
-			return ProcessorProfiles == null ||
-				(ProfileMgr == null && winKeyboard.InputProcessorProfile.Hkl == IntPtr.Zero);
-		}
-
-		private void SwitchKeyboard(WinKeyboardDescription winKeyboard)
-		{
-			if (_fSwitchingKeyboards)
-				return;
-
-			_fSwitchingKeyboards = true;
-			try
-			{
-				KeyboardController.Instance.ActiveKeyboard = ActivateKeyboard(winKeyboard);
-				if (Form.ActiveForm != null)
-				{
-					// If we activate a keyboard while a particular Form is active, we want to know about
-					// input language change calls for that form. The previous -= may help make sure we
-					// don't get multiple hookups.
-					Form.ActiveForm.InputLanguageChanged -= ActiveFormOnInputLanguageChanged;
-					Form.ActiveForm.InputLanguageChanged += ActiveFormOnInputLanguageChanged;
-				}
-
-				// If we have a TIP (TSF Input Processor) we don't have a handle. But we do the
-				// keyboard switching through TSF so we don't need the workaround below.
-				if (!UseWindowsApiForKeyboardSwitching(winKeyboard))
-					return;
-
-				_expectedKeyboard = winKeyboard;
-				// The following lines help to work around a Windows bug (happens at least on
-				// XP-SP3): When you set the current input language (by any method), if there is more
-				// than one loaded input language associated with that same culture, Windows may
-				// initially go along with your request, and even respond to an immediate query of
-				// the current input language with the answer you expect.  However, within a fraction
-				// of a second, it often takes the initiative to again change the input language to
-				// the _other_ input language having that same culture. We check that the proper
-				// input language gets set by enabling a timer so that we can re-set the input
-				// language if necessary.
-				_fSwitchedLanguages = true;
-				// stop timer first so that the 0.5s interval restarts.
-				_timer.Stop();
-				_timer.Start();
-			}
-			finally
-			{
-				_fSwitchingKeyboards = false;
 			}
 		}
-
-		private WinKeyboardDescription ActivateKeyboard(WinKeyboardDescription winKeyboard)
-		{
-			try
-			{
-				if (UseWindowsApiForKeyboardSwitching(winKeyboard))
-				{
-					// Win XP with regular keyboard, or TSF disabled
-					Win32.ActivateKeyboardLayout(new HandleRef(this, winKeyboard.InputLanguage.Handle), 0);
-					return winKeyboard;
-				}
-
-				var profile = winKeyboard.InputProcessorProfile;
-				if ((profile.Flags & TfIppFlags.Enabled) == 0)
-					return winKeyboard;
-
-				ProcessorProfiles.ChangeCurrentLanguage(profile.LangId);
-				if (ProfileMgr == null)
-				{
-					// Win XP with TIP (TSF Input Processor)
-					ProcessorProfiles.ActivateLanguageProfile(ref profile.ClsId, profile.LangId,
-						ref profile.GuidProfile);
-				}
-				else
-				{
-					// Windows >= Vista with either TIP or regular keyboard
-					ProfileMgr.ActivateProfile(profile.ProfileType, profile.LangId,
-						ref profile.ClsId, ref profile.GuidProfile, profile.Hkl,
-						TfIppMf.DontCareCurrentInputLanguage);
-				}
-			}
-			catch (ArgumentException)
-			{
-				// throws exception for non-supported culture, though seems to set it OK.
-			}
-			catch (COMException e)
-			{
-				var profile = winKeyboard.InputProcessorProfile;
-				var msg = string.Format("Got COM exception trying to activate IM:" + Environment.NewLine +
-					"LangId={0}, clsId={1}, hkl={2}, guidProfile={3}, flags={4}, type={5}, catId={6}",
-					profile.LangId, profile.ClsId, profile.Hkl, profile.GuidProfile, profile.Flags, profile.ProfileType, profile.CatId);
-				throw new ApplicationException(msg, e);
-			}
-			return winKeyboard;
-		}
-
-		/// <summary>
-		/// Save the state of the conversion and sentence mode for the current IME
-		/// so that we can restore it later.
-		/// </summary>
-		private void SaveImeConversionStatus(WinKeyboardDescription winKeyboard)
-		{
-			if (winKeyboard == null)
-				return;
-
-			var windowHandle = new HandleRef(this,
-				winKeyboard.WindowHandle != IntPtr.Zero ? winKeyboard.WindowHandle : Win32.GetFocus());
-			var contextPtr = Win32.ImmGetContext(windowHandle);
-			if (contextPtr == IntPtr.Zero)
-				return;
-
-			var contextHandle = new HandleRef(this, contextPtr);
-			int conversionMode;
-			int sentenceMode;
-			Win32.ImmGetConversionStatus(contextHandle, out conversionMode, out sentenceMode);
-			winKeyboard.ConversionMode = conversionMode;
-			winKeyboard.SentenceMode = sentenceMode;
-			Win32.ImmReleaseContext(windowHandle, contextHandle);
-		}
-
-		/// <summary>
-		/// Restore the conversion and sentence mode to the states they had last time
-		/// we activated this keyboard (unless we never activated this keyboard since the app
-		/// got started, in which case we use sensible default values).
-		/// </summary>
-		private void RestoreImeConversionStatus(IKeyboardDefinition keyboard)
-		{
-			var winKeyboard = keyboard as WinKeyboardDescription;
-			if (winKeyboard == null)
-				return;
-
-			// Restore the state of the new keyboard to the previous value. If we don't do
-			// that e.g. in Chinese IME the input mode will toggle between English and
-			// Chinese (LT-7487 et al).
-			var windowPtr = winKeyboard.WindowHandle != IntPtr.Zero ? winKeyboard.WindowHandle : Win32.GetFocus();
-			var windowHandle = new HandleRef(this, windowPtr);
-
-			// NOTE: Windows uses the same context for all windows of the current thread, so it
-			// doesn't really matter which window handle we pass.
-			var contextPtr = Win32.ImmGetContext(windowHandle);
-			if (contextPtr == IntPtr.Zero)
-				return;
-
-			// NOTE: Chinese Pinyin IME allows to switch between Chinese and Western punctuation.
-			// This can be selected in both Native and Alphanumeric conversion mode. However,
-			// when setting the value the punctuation setting doesn't get restored in Alphanumeric
-			// conversion mode, not matter what I try. I guess that is because Chinese punctuation
-			// doesn't really make sense with Latin characters.
-			var contextHandle = new HandleRef(this, contextPtr);
-			Win32.ImmSetConversionStatus(contextHandle, winKeyboard.ConversionMode, winKeyboard.SentenceMode);
-			Win32.ImmReleaseContext(windowHandle, contextHandle);
-			winKeyboard.WindowHandle = windowPtr;
-		}
-
-		private void ActiveFormOnInputLanguageChanged(object sender, InputLanguageChangedEventArgs inputLanguageChangedEventArgs)
-		{
-			RestoreImeConversionStatus(GetKeyboardDescription(inputLanguageChangedEventArgs.InputLanguage.Interface()));
-		}
-
-		private static bool InputProcessorProfilesEqual(TfInputProcessorProfile profile1, TfInputProcessorProfile profile2)
-		{
-			// Don't compare Flags - they can be different and it's still the same profile
-			return profile1.ProfileType == profile2.ProfileType &&
-				profile1.LangId == profile2.LangId &&
-				profile1.ClsId == profile2.ClsId &&
-				profile1.GuidProfile == profile2.GuidProfile &&
-				profile1.CatId == profile2.CatId &&
-				profile1.HklSubstitute == profile2.HklSubstitute &&
-				profile1.Caps == profile2.Caps &&
-				profile1.Hkl == profile2.Hkl;
-		}
-
+		
 		#region IKeyboardRetrievingAdaptor Members
 
 		/// <summary>
@@ -726,38 +370,15 @@ namespace SIL.Windows.Forms.Keyboarding.Windows
 			}
 		}
 
-		public bool IsApplicable
-		{
-			get { return true; }
-		}
-
-		public IKeyboardSwitchingAdaptor SwitchingAdaptor
-		{
-			get { return this; }
-		}
-
-		[SuppressMessage("Gendarme.Rules.Correctness", "EnsureLocalDisposalRule",
-			Justification = "m_Timer gets disposed in Close() which gets called from KeyboardControllerImpl.Dispose")]
 		public void Initialize()
 		{
-			_timer = new Timer { Interval = 500 };
-			_timer.Tick += OnTimerTick;
+			UpdateAvailableKeyboards();
 
-			GetInputMethods();
-
-			// Form.ActiveForm can be null when running unit tests
-			if (Form.ActiveForm != null)
-				Form.ActiveForm.InputLanguageChanged += ActiveFormOnInputLanguageChanged;
-		}
-
-		public void UpdateAvailableKeyboards()
-		{
-			GetInputMethods();
 		}
 
 		private KeyboardDescription GetKeyboardForInputLanguage(IInputLanguage inputLanguage)
 		{
-			return GetKeyboardDescription(inputLanguage);
+			return WinKeyboardUtils.GetKeyboardDescription(inputLanguage);
 		}
 
 		/// <summary>
@@ -773,21 +394,8 @@ namespace SIL.Windows.Forms.Keyboarding.Windows
 			string locale = parts[0];
 			string layout = parts.Length > 1 ? parts[1] : string.Empty;
 
-			IInputLanguage inputLanguage;
 			string cultureName;
-			try
-			{
-				var ci = new CultureInfo(locale);
-				cultureName = ci.DisplayName;
-				inputLanguage = new InputLanguageWrapper(ci, IntPtr.Zero, layout);
-			}
-			catch (CultureNotFoundException)
-			{
-				// ignore if we can't find a culture (this can happen e.g. when a language gets
-				// removed that was previously assigned to a WS) - see LT-15333
-				inputLanguage = null;
-				cultureName = "[Unknown Language]";
-			}
+			var inputLanguage = WinKeyboardUtils.GetInputLanguage(locale, layout, out cultureName);
 
 			return new WinKeyboardDescription(id, GetDisplayName(layout, cultureName), layout, locale, false, inputLanguage, this,
 				GetDisplayName(layout, cultureName), new TfInputProcessorProfile());
@@ -815,58 +423,6 @@ namespace SIL.Windows.Forms.Keyboarding.Windows
 		public bool IsSecondaryKeyboardSetupApplication
 		{
 			get { return false; }
-		}
-
-		#endregion
-
-		#region IKeyboardSwitchingAdaptor Members
-
-		public bool ActivateKeyboard(KeyboardDescription keyboard)
-		{
-			CheckDisposed();
-			SwitchKeyboard((WinKeyboardDescription) keyboard);
-			return true;
-		}
-
-		public void DeactivateKeyboard(KeyboardDescription keyboard)
-		{
-			CheckDisposed();
-			SaveImeConversionStatus((WinKeyboardDescription) keyboard);
-		}
-
-		/// <summary>
-		/// Gets the default keyboard of the system.
-		/// </summary>
-		public KeyboardDescription DefaultKeyboard
-		{
-			get
-			{
-				CheckDisposed();
-				return GetKeyboardDescription(InputLanguage.DefaultInputLanguage.Interface());
-			}
-		}
-
-		/// <summary>
-		///  Gets the currently active keyboard.
-		/// </summary>
-		public KeyboardDescription ActiveKeyboard
-		{
-			get
-			{
-				if (ProfileMgr != null)
-				{
-					var profile = ProfileMgr.GetActiveProfile(ref Guids.Consts.TfcatTipKeyboard);
-					return Keyboard.Controller.AvailableKeyboards.OfType<WinKeyboardDescription>()
-						.FirstOrDefault(winKeybd => InputProcessorProfilesEqual(profile, winKeybd.InputProcessorProfile)) ??
-						KeyboardController.NullKeyboard;
-				}
-
-				// Probably Windows XP where we don't have ProfileMgr
-				var lang = ProcessorProfiles.GetCurrentLanguage();
-				return Keyboard.Controller.AvailableKeyboards.OfType<WinKeyboardDescription>()
-					.FirstOrDefault(winKeybd => winKeybd.InputProcessorProfile.LangId == lang) ??
-					KeyboardController.NullKeyboard;
-			}
 		}
 
 		#endregion
@@ -948,13 +504,6 @@ namespace SIL.Windows.Forms.Keyboarding.Windows
 
 			if (disposing)
 			{
-				// Dispose managed resources here.
-				if (_timer != null)
-				{
-					_timer.Dispose();
-					_timer = null;
-				}
-
 				if (_profileNotifySinkCookie > 0)
 				{
 					if (TfSource != null)
